+++ router_test.go
@@ -68,6 +68,8 @@ func (suite *RouterTestSuite) TestCleanStaticPath() {
 	suite.Equal("resources/lang/en-US/locale.json", cleanStaticPath("resources", "/lang/en-US/locale.json"))
 	suite.Equal("resources/lang/en-US/index.html", cleanStaticPath("resources", "lang/en-US"))
 	suite.Equal("resources/lang/en-US/index.html", cleanStaticPath("resources", "lang/en-US/"))
+	suite.Equal("resources/lang/index.html", cleanStaticPath("resources", "lang"))
+	suite.Equal("resources/lang/index.html", cleanStaticPath("resources", "lang/"))
 }
 
 func (suite *RouterTestSuite) TestStaticHandler() {