+++ docs_src/src/guide/changelog.md
@@ -20,3 +20,4 @@
 - `goyave.Response` now implements `http.ResponseWriter`.
     - All writing functions can now return an error.
 - Added the [`NativeHandler`](./basics/routing#native-handlers) compatibility layer.
+- Fixed a bug preventing the static resources handler to find `index.html` if a directory with a depth of one was requested without a trailing slash.